requirements.txt
Django
djangorestframework
pygithub
django-cors-headers
mysqlclient
django-crontab